refactor(ocr): hoist shared empty-structured detection + recovery merge to BaseOCR
Phase 4. _is_empty_structured was byte-identical across the OpenAI and Vertex
providers, and the final merge loop of _recover_empty_structured (apply recovered
free-form text back into the empty results, tag structured_fallback='free_form')
was byte-identical too. Move both to BaseOCR (_is_empty_structured static +
_apply_recovered); both providers delegate. The provider-specific part — the
free-form re-OCR setup (OpenAI restores its structured agent in try/finally; Vertex
filters kwargs) — stays local. Zero behavior change (identical code moved): 821 tests
pass incl. the recovery suite, and live 2-page OCR is unchanged.
