+++ doc2mark/ocr/vertex_ai.py
@@ -500,18 +500,6 @@ class VertexAIOCR(BaseOCR):
             prompts.append(base)
         return prompts
 
-    @staticmethod
-    def _is_empty_structured(result: OCRResult) -> bool:
-        """A structured result with no usable content (some models/images cannot
-        fill the json_schema and return an empty OCRPage)."""
-        if result.text and result.text.strip():
-            return False
-        doc = result.document
-        if doc is None:
-            return True
-        raw = doc.raw
-        return not (raw.text.strip() or raw.tables or raw.fields)
-
     def _recover_empty_structured(
         self, results: List[OCRResult], images: List[bytes], **kwargs
     ) -> List[OCRResult]:
@@ -538,26 +526,7 @@ class VertexAIOCR(BaseOCR):
         recovered = self._batch_process_with_vision_agent(
             [images[i] for i in empty_idx], **fk
         )
-
-        for j, i in enumerate(empty_idx):
-            text = (recovered[j].text or "").strip()
-            if not text:
-                continue
-            doc = results[i].document
-            if doc is not None:
-                doc.raw.text = text
-            else:
-                doc = OCRPage(raw=RawExtraction(text=text))
-            meta = dict(results[i].metadata or {})
-            meta["structured_fallback"] = "free_form"
-            results[i] = OCRResult(
-                text=text,
-                confidence=results[i].confidence,
-                language=results[i].language,
-                metadata=meta,
-                document=doc,
-            )
-        return results
+        return self._apply_recovered(results, empty_idx, recovered)
 
     def _batch_process_with_vision_agent(
         self, images: List[bytes], **kwargs
